move ParamEnv methods from `ty/util` to `ty/mod`
